Don't try to kill own process (0) - only real child processes.

+++ src/camera.c
@@ -38,7 +38,7 @@ void TerminateMonitor( int signo )
 
 void PingCameras( int signo )
   {
-  Notice("Recived SIGHUP.  Will try to reactivate non-functional cameras.");
+  Notice("Recieved SIGHUP.  Will try to reactivate non-functional cameras.");
   if( glob_conf==NULL )
     return;
   if( glob_conf->cameras==NULL )
@@ -227,35 +227,38 @@ pid_t LaunchCapture( _CONFIG* config, _CAMERA* cam )
 
   Notice( "LaunchCapture on %s", cam->nickName );
 
-  /* do we already have one?  if so, kill it first. */
-  if( ProcessExistsAndIsMine( cam->childProcess )==0 )
+  if( cam->childProcess )
     {
-    Notice("LaunchCapture( %s ) -- process already exists", cam->nickName );
-    int err = 0;
-
-    err = kill( cam->childProcess, SIGHUP );
-    if( err==0 )
+    /* do we already have one?  if so, kill it first. */
+    if( ProcessExistsAndIsMine( cam->childProcess )==0 )
       {
-      sleep(1);
-      if( ProcessExistsAndIsMine( cam->childProcess )==0 )
-        { /* needs a more brutal kill */
-        Warning( "LaunchCapture( %s ) -- kill -1 not enough", cam->nickName );
-        err = kill( cam->childProcess, SIGKILL );
+      Notice("LaunchCapture( %s ) -- process already exists", cam->nickName );
+      int err = 0;
+
+      err = kill( cam->childProcess, SIGHUP );
+      if( err==0 )
+        {
         sleep(1);
         if( ProcessExistsAndIsMine( cam->childProcess )==0 )
-          {
-          Warning( "LaunchCapture( %s ) -- kill -9 not enough!!", cam->nickName );
+          { /* needs a more brutal kill */
+          Warning( "LaunchCapture( %s ) -- kill -1 not enough", cam->nickName );
+          err = kill( cam->childProcess, SIGKILL );
+          sleep(1);
+          if( ProcessExistsAndIsMine( cam->childProcess )==0 )
+            {
+            Warning( "LaunchCapture( %s ) -- kill -9 not enough!!", cam->nickName );
+            }
           }
+        }
+      else
+        {
+        Warning( "LaunchCapture( %s ) -- failed to kill -HUP existing child", cam->nickName );
         }
       }
     else
       {
-      Warning( "LaunchCapture( %s ) -- failed to kill -HUP existing child", cam->nickName );
+      Notice( "LaunchCapture on %s - no legacy process.", cam->nickName );
       }
-    }
-  else
-    {
-    Notice( "LaunchCapture on %s - no legacy process.", cam->nickName );
     }
 
   cam->childProcess = 0;
